Revert "Revert "DescriptorMerger""
This reverts commit 69e7f87ca964d2c7e5f089d8eff12c465cc0b245.
