+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/handlers/v2/DescriptorMerger.java
@@ -3,18 +3,19 @@ package com.sap.cloud.lm.sl.mta.handlers.v2;
 import com.sap.cloud.lm.sl.mta.mergers.v2.ExtensionDescriptorMerger;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionDescriptor;
 
-public class DescriptorMerger extends com.sap.cloud.lm.sl.mta.handlers.v1.DescriptorMerger {
+public class DescriptorMerger {
+
+    protected final DescriptorHandler handler;
 
     public DescriptorMerger() {
-        super(new DescriptorHandler());
+        this(new DescriptorHandler());
     }
 
     public DescriptorMerger(DescriptorHandler handler) {
-        super(handler);
+        this.handler = handler;
     }
 
-    @Override
-    public ExtensionDescriptorMerger getExtensionDescriptorMerger(com.sap.cloud.lm.sl.mta.model.v1.ExtensionDescriptor extension) {
+    public ExtensionDescriptorMerger getExtensionDescriptorMerger(ExtensionDescriptor extension) {
         return new ExtensionDescriptorMerger((ExtensionDescriptor) extension, (DescriptorHandler) handler);
     }
